+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -1,5 +1,11 @@
 # TestCommon Release notes
 
+## Version 1.3.0
+
+- Implemented `FunctionAppHostManager.RestartHostIfChanges(IEnumerable<KeyValuePair<string, string>> environmentVariables)` that only restarts the function app if process environment variables has changed.
+- Extended control of starting an Azure Function App through use of the `FunctionAppHostManager`:
+  - Specify which host log message to await during startup using `FunctionAppHostSettings.HostStartedEvent`.
+
 ## Version 1.2.0
 
 - Extended `IntegrationTestConfiguration` to support using Event Hub.